split subscribers to another file
